// apps/v2/src/app/services/sdk/types.gen.ts
export type MovieCommentDto = {
  _id: string;
  /**
   * Rating
   */
  rating: number;
  userId: string;
  userName: string;
  movieId: string;
  content: string;
};

export type CreateMovieCommentDto = {
  /**
   * Rating
   */
  rating: number;
  movieId: string;
  content: string;
};